Fix [Artifacts] Add a confirmation pop-up to details actions menu
