+++ src/components/Documents/documents.util.js
@@ -138,7 +138,8 @@ export const generateActionsMenu = (
         label: 'Add a tag',
         hidden: isDetailsPopUp,
         icon: <TagIcon />,
-        onClick: handleAddTag
+        onClick: handleAddTag,
+        allowLeaveWarning: true
       },
       {
         label: 'Download',
@@ -189,7 +190,8 @@ export const generateActionsMenu = (
             filters,
             refreshArtifacts,
             refreshAfterDeleteCallback
-          })
+          }),
+        allowLeaveWarning: true
       },
       {
         label: 'Delete all versions',
@@ -214,7 +216,8 @@ export const generateActionsMenu = (
                 true
               )
             }
-          )
+          ),
+        allowLeaveWarning: true
       }
     ],
     [
