+++ src/components/Files/files.util.js
@@ -159,7 +159,8 @@ export const generateActionsMenu = (
         label: 'Add a tag',
         hidden: isDetailsPopUp,
         icon: <TagIcon />,
-        onClick: handleAddTag
+        onClick: handleAddTag,
+        allowLeaveWarning: true
       },
       {
         label: 'Download',
@@ -210,7 +211,8 @@ export const generateActionsMenu = (
             filters,
             refreshArtifacts,
             refreshAfterDeleteCallback
-          })
+          }),
+        allowLeaveWarning: true
       },
       {
         label: 'Delete all versions',
@@ -235,7 +237,8 @@ export const generateActionsMenu = (
                 true
               )
             }
-          )
+          ),
+        allowLeaveWarning: true
       }
     ],
     [
