+++ src/components/ModelsPage/Models/models.util.js
@@ -233,7 +233,8 @@ export const generateActionsMenu = (
         label: 'Add a tag',
         hidden: isDetailsPopUp,
         icon: <TagIcon />,
-        onClick: handleAddTag
+        onClick: handleAddTag,
+        allowLeaveWarning: true
       },
       {
         label: 'Download',
@@ -292,7 +293,8 @@ export const generateActionsMenu = (
                 MODEL_TYPE
               )
             }
-          )
+          ),
+        allowLeaveWarning: true
       },
       {
         label: 'Delete all versions',
@@ -317,7 +319,8 @@ export const generateActionsMenu = (
                 true
               )
             }
-          )
+          ),
+        allowLeaveWarning: true
       }
     ],
     [
